Update dependency versions and refactor project architecture.
