+++ .github/workflows/pack.yml
@@ -1,0 +1,30 @@
+name: build_main_binary
+on: 
+  push:
+    branches:
+      - publish
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    permissions:
+      id-token: write
+    steps:
+      - name: Checkout Code
+        uses: actions/checkout@v4.1.1
+      - name: Setup Python
+        uses: actions/setup-python@v5.0.0
+        with:
+          python-version: 3.11
+          architecture: x64
+      - name: Install Pipenv
+        run: pip install pipenv
+      - name: Install Dependencies
+        run: pipenv install --dev
+      - name: Build PyPI package
+        run: pipenv run python -m build
+      - name: Publish to PyPI
+        env:
+          PYPI_TOKEN: ${{ secrets.PYPI_TOKEN }}
+        shell: cmd
+        run: pipenv run twine upload ./dist/*.whl -r pypi -u "__token__" -p "%PYPI_TOKEN%"
+        continue-on-error: true
