--- About_v3next.md
@@ -1,3 +1,0 @@
-# About V3Next
-
-[中文](about_v3next_cn.md)
