--- about_v3next_cn.md
@@ -1,36 +1,0 @@
-# 关于MCSMT V3 Next的重大更新概况及提醒
-
-## 发生什么事了？
-
-我们推出了 我的世界客户端管理工具(MCSMT) V3 Next，这是我们对 MCSMT V3 的重大更新。
-该更新包含了很多不具有确定性的特性，包括：
-- 增加了用户界面（基于Electron）
-- 增加了对在线下载Curseforge上的Mod的支持
-
-我们暂时没有在任何平台测试过本次更新，所以若有人愿意进行测试，欢迎。请在Github上提交issue。
-
-## 我该如何使用MCSMT V3 Next？
-
-我们目前没有发布任何预编译版本，你需要从Github下载或克隆源代码并在本机编译。
-
-### 一、生成核心组件的可执行文件
-
-安装并进入Pipenv环境，进入`src`目录，然后使用你喜欢的任何编译器来编译`mcsmtGo.py`并得到可执行文件`mcsmtGo`或`mcsmtGo.exe`。一般来说，这将自动包含运行所需要的所有依赖项。
-
-### 二、编译用户界面
-
-用户界面和核心组件是完全独立的，所以你需要单独编译它。
-你可以用你喜欢的任何方式编译这个Electron应用。
-目录是：`electron_ui`
-可执行文件应该名为`mcsmtUi`或`mcsmtUi.exe`，否则，重命名它。
-
-### 三、编译连接件
-
-因为需要同时启动用户界面和核心组件，所以为了方便，我们提供了一个连接件的源代码，它能够同时运行用户界面和核心组件。你只需要使用任何C/C++编译器编译`starter.cpp`，并得到可执行文件。
-
-### 四、使用MCSMT V3 Next
-
-现在，把连接件、核心组件和用户界面的可执行文件都放在同一个目录下，然后运行连接件。
-
-现在你应该能看到MCSMT V3 Next的界面了。
-
