+++ src/filesync/__main__.py
@@ -1,0 +1,14 @@
+print(
+f'''
+Elemental File Sync v0.3.1
+Just like Git, but without version control.
+
+You shouldn't use this command.
+Use `<python> -m filesync.<someModule> [someArgument]` instead.
+There's the module list:
+    - client
+        Runs the client mode of EFS to clone files from server. See GitHub Wiki for more information.
+    - server
+        Runs the server mode of EFS to generate index file. See GitHub Wiki for more information.
+'''
+)
