--- src/mcsmt/__init__.py
@@ -1,7 +1,0 @@
-from . import client, mixed, server
-
-__all__ = [
-    'client',
-   'mixed',
-   'server'
-]
