--- src/mcsmt/__main__.py
@@ -1,2 +1,0 @@
-from .starter import main
-main()
