--- src/mcsmt/client/__init__.py
@@ -1,5 +1,0 @@
-from . import main
-
-__all__ = [
-    "main"
-]
