--- src/mcsmt/info.py
@@ -1,5 +1,0 @@
-mcsmt_modules = {
-    "client": ["main"],
-    "mixed": ["down", "extdown", "ezdns", "moddown"],
-    "server": ["gen", "genconf", "wdgen"]
-}
