--- src/mcsmt/mixed/__init__.py
@@ -1,8 +1,0 @@
-from . import down, extdown, ezdns, moddown
-
-__all__ = [
-    "down",
-    "extdown",
-    "ezdns",
-    "moddown",
-]
