--- src/mcsmt/mixed/about_ModLoader.txt
@@ -1,15 +1,0 @@
-ModLoaderType
-
-Possible enum values:
-
-0=Any
-
-1=Forge
-
-2=Cauldron
-
-3=LiteLoader
-
-4=Fabric
-
-5=Quilt
