--- src/mcsmt/server/__init__.py
@@ -1,7 +1,0 @@
-from . import gen, genconf, wdgen
-
-__all__ = [
-    "gen",
-    "genconf",
-    "wdgen",
-]
