--- src/mcsmtGo.py
@@ -1,2 +1,0 @@
-from mcsmt.starter import main
-main()
